+++ front/components/assistant/AssistantBrowser.tsx
